+++ Observer/Validate/Frontend.php
@@ -10,30 +10,22 @@ declare(strict_types=1);
 
 namespace PixelOpen\CloudflareTurnstile\Observer\Validate;
 
-use Magento\Contact\Controller\Index\Post as ContactPost;
-use Magento\Customer\Controller\Account\CreatePost;
 use Magento\Customer\Controller\Ajax\Login as AjaxLoginPost;
 use Magento\Customer\Model\Session as CustomerSession;
 use Magento\Framework\App\ActionInterface;
-use Magento\Framework\App\Request\DataPersistorInterface;
 use Magento\Framework\App\Request\Http as Request;
 use Magento\Framework\App\Response\Http as Response;
 use Magento\Framework\Message\ManagerInterface;
 use Magento\Framework\Phrase;
 use Magento\Framework\Serialize\Serializer\Json;
-use Magento\Framework\Session\Generic;
-use Magento\Review\Controller\Product\Post as ReviewPost;
 use PixelOpen\CloudflareTurnstile\Helper\Config;
+use PixelOpen\CloudflareTurnstile\Model\PersistorInterface;
 use PixelOpen\CloudflareTurnstile\Model\Validator;
 use PixelOpen\CloudflareTurnstile\Observer\Validate;
 
 class Frontend extends Validate
 {
-    protected DataPersistorInterface $dataPersistor;
-
     protected CustomerSession $customerSession;
-
-    protected Generic $reviewSession;
 
     /**
      * @param ManagerInterface $messageManager
@@ -41,9 +33,8 @@ class Frontend extends Validate
      * @param Validator $validator
      * @param Json $json
      * @param Config $config
-     * @param DataPersistorInterface $dataPersistor
      * @param CustomerSession $customerSession
-     * @param Generic $reviewSession
+     * @param PersistorInterface|null $persistor
      * @param array $data
      */
     public function __construct(
@@ -52,17 +43,15 @@ class Frontend extends Validate
         Validator $validator,
         Json $json,
         Config $config,
-        DataPersistorInterface $dataPersistor,
         CustomerSession $customerSession,
-        Generic $reviewSession,
+        ?PersistorInterface $persistor = null,
         array $data = []
     ) {
-        $this->dataPersistor = $dataPersistor;
         $this->customerSession = $customerSession;
-        $this->reviewSession = $reviewSession;
 
-        parent::__construct($messageManager, $response, $validator, $json, $config, $data);
+        parent::__construct($messageManager, $response, $validator, $json, $config, $persistor, $data);
     }
+
 
     /**
      * Can validate action
@@ -99,28 +88,6 @@ class Frontend extends Validate
     public function isEnabled(): bool
     {
         return $this->config->isEnabledOnFront();
-    }
-
-    /**
-     * Persist data
-     *
-     * @param Request $request
-     * @param ActionInterface $action
-     * @return bool
-     */
-    public function persist(Request $request, ActionInterface $action): bool
-    {
-        if ($action instanceof ContactPost) {
-            $this->dataPersistor->set('contact_us', $request->getParams());
-        }
-        if ($action instanceof CreatePost) {
-            $this->customerSession->setCustomerFormData($request->getParams());
-        }
-        if ($action instanceof ReviewPost) {
-            $this->reviewSession->setFormData($request->getParams());
-        }
-
-        return true;
     }
 
     /**